improve logging practices
Logging and then returning an error produces multiple messages, so
change a few spots to just return and let the caller do the logging.

Tighten up the language in a few log messages so they are shorter.

Wait until an action is done or failed to log, so we have "did action"
or "failed action" messages but no "doing action" messages. This
ensures that all of the information we need to have is in 1 log
message, instead of split across several.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>