Fix initialiation of KubeClient
The Client added by KubeBuilder reads resource values from a cache
but writes to the API server. Since the code that creates resources
for the metal3 deployment creates resources just before using them,
we are taking the prudent approach of adding a regualr k8s client
that reads and writes to the API Server.
Some more experimentation is required before we can confidently use
the Client to create and update the metal3 Deployment object and the
resources that it depends on.