WIP: controllers: "Unsupported", etc. condition messaging
"Operator is non-functional" sounds a bit frightening.  This commit
pivots to say "Operator has no role on platform {platform-name}" in
the cases where we determine we are not needed.

I'm also removing ReasonEmpty, because it's always nice to say
*something* about why we picked the state we picked.  In the
ReasonInvalidConfiguration and ReasonDeployTimedOut cases, that makes
me wonder if we really intend to be Available then.  Maybe we should
just echo whatever we're setting for Degraded in that case?  FIXME
until we decide.