Cache OS images on host volume
Currently OS images are cached on a volume that is internal to the
metal3 Pod. To allow them to be shared with a second-level cache, store
them on the host at /var/lib/metal3/images.

This will allow us to cache the images on each master node so that even
after failover of the Metal3 Pod we will never need to re-download the
image from outside the cluster.