Add a second-level OS image cache
Add a cache for OS images accessible on port 6181 on every master node.
This cache downloads the images from the first-level cache in the metal3
Pod (port 6180). The images are on shared storage, so even when the
metal3 Pod moves, we'll never need to download them from the Internet
again.