Allow metal3 Deployment to be updated
In order to avoid fighting with the MAO, we temporarily avoid updating
any existing metal3 Deployment we find. However, this means that
Deployments created by the CBO cannot be updated, and we cannot
gracefully handle any errors that occur after initially creating the
Deployment.

Allow fuller testing of the CBO by only refusing to update the
Deployment if it is annotated as being managed by the MAO.

In future we will take over management of Deployments created by the MAO
and substitute our own annotation.