Store expected generations in CR
Previously we stored the expected generations for resources being
managed by the CBO only in memory. This was a holdover from the MAO,
which doesn't use the operators.openshift.io API and therefore could
only store the data in memory (see openshift/machine-api-operator#541).

The CBO uses the operators API, and can therefore store the expected
generations in the Status. This means that resources won't be rewritten
unnecessarily just because the CBO restarts.