Add a Service for the stateful parts of the metal3 Pod
The metal3 Pod contains the (ephemeral) internal state of Metal³, and is
able to float between different master nodes. In future, we want to
deploy components outside the Pod that nevertheless interact with the
internal state. To allow them to locate it, create a Service that points
to the Pod.

Currently the httpd endpoint is the only one exposed, but in future the
Ironic DB may be also.