Use hostIP for ironic containers when provisionig network disabled
This change makes the ironic related services use the hostIP when the
provisioningIP is blank. This removes the requirement for a
provisioningIP in Ironic, but additional work is needed to fix the image
URL's we pass to baremetal-operator.

As we are using an existing IP on an interface, we no longer have a
requirement on the static IP manager when the provisioningIP is blank,
so we do not start the containers in this case.

This also changes baremetal operator to communicate with the ironic
containers on localhost, as they are in the same pod.