Don't do rolling updates of metal3 Deployment
Up to now we have relied on the ironic-static-ip-manager containers
managing the Provisioning IP to ensure that only one metal3 Pod is
started at a time. However, now that we support configurations without
such an IP, we should take other measures to ensure that we don't create
multiple inconsistent copies of the state (e.g. ironic database).

Changing the Deployment Strategy from the default rolling update to
recreate means that upon any changes to the Pod template, the existing
ReplicaSet will be deleted before an updated one is created.