Serve virtual media ISOs over TLS
Especially in the live ISO case, the virtual media images may contain
sensitive information. Now that we've established that Redfish hardware
does not verify TLS certificates, we can use our self-signed certificate
to protect the communication with the BMC.

A knob to turn this feature off is provided for hardware that diverges.

Requires: https://github.com/openshift/ironic-image/pull/218