Fix image publish address and use ClusterIP service
The image publish address needs to be set on the command line, not in
environment variables, so it was using the default (127.0.0.1:8084).

The original idea behind using a NodePort Service was that it would be
accessible to the BMCs. In practice it turns out we don't have a
requirement for that yet - currently ironic is the only thing
downloading images, and it caches them to serve up to the BMC. So for
now downgrade this to a ClusterIP service (one of which is created by
the NodePort Service anyway). We can always change it to NodePort if
this becomes a requirement.

The ClusterIP service can be referenced using the internal DNS name. To
simplify URLs, expose the service on port 80 on the Pod's cluster IP.