Replace Pre-provisioning downloaders with machine-os-images
The machine-os-images container supplies the latest ISO and PXE images
for pre-provisioning from within the release payload. Use this instead
of downloading preprovisioning images or using the RHEL IPA. Extract the
images to the same location (and with the same names) in the image cache
volume that we were using previously for IPA images. This change is
transparent to the tasks using the images.

Continue to run the second-level image cache service and download the
Provisioning OS QCOW (if specified). Existing clusters installed prior
to 4.10 will contain MachineSets that still rely on the QCOW for
provisioning new Machines.