Create image-customization-controller when metal3 Pod not up
Even if there is no metal3 Pod up and therefore we cannot find its IP
address, we still want to create the image-customization-controller with
an empty ironic IP address.

This is documented in a comment at the site where getPodHostIP() is
called, but not in the implementation, so it was inadvertently broken
by #225.