Make IP options consistent between virtualmedia and PXE
The IP_OPTIONS environment variable in machine-os-images configures the
kernel command line in the ISO image. The IRONIC_KERNEL_PARAMS
environment variable in ironic-conductor/ironic-inspector configures the
kernel command line in the PXE image. The latter was changed to be based
on the address family of the provisioning network in
355f862c0ba482f40e26daafc385dbbc89ffaed5.

The PXE and ISO methods should be consistent, so also use the
provisioning network's address family to determine the IP options for
the ISO. An exception is if the virtualMediaViaExternalNetwork option is
set: in this case hosts booted from ISOs will contact ironic via the
external network even if the provisioning network exists.