Run a proxy for ironic when provisioning network is disabled
Without a provisioning network, ICC uses IP address of the control plane
node with the metal3 pod. If the pod relocates, the ramdisk will lose
access to Ironic. To solve this issue, run a simple HTTP(s) reverse proxy
as a DaemonSet on all control plane nodes.

This commit is incomplete in a sense that it does not pass the right IP
address to ICC. It will be addressed in a follow-up.