OCPBUGS-651: ignore metal3 pods that are terminating
When a master abruptly terminates, a replacement will be found and a new
copy of the metal3 pod will be created. However, the old copy will be
left Terminating until the master comes back up or the pod is
force-deleted. This currently prevents the ICC deployment from being
updated since it expects only one copy of the pod. To work around this
problem, ignore pods with DeletionTimestamp when calculating the current
metal3 pod IP address.