Fix compilation failures due to new dependencies
These all mostly just add a context to function calls.
The webhook-related functions have also added a cache parameter, but
with earlier versions this was just not being used. Because of this it
is safe to continue to create an empty cache each call, but in the
future it may be useful to use correctly.