Revert "proxy: pass ipv6 address to container"
First and foremost, it made the proxy refer to itself, which only
happens to work when the API VIP is on the same master as the metal3
pod. Proxy must use the real pod IP.

Then, the UPSTREAM_*_V6 variables do not exist in ironic-image, setting
them does nothing. Passing several upstream IPs does not really make
much sense: we expect that Ironic is available on any of the valid IPs,
and we can make the proxy contact any of them.

This reverts commit ba7403e57e344e615d375af5cb7bfec5d56f03e9.