Avoid ironic-proxy when setting external url
The external url is the url the BMC will use to retrieve images from the
httpd container inside metal3 in dual stack scenarios.

Each time the metal3 pod is started, we have to find any existing nodes,
and rebuild the Ironic database.  During this process, we also tell each
BMC where to get images.  This means that httpd doesn't have to be
proxied, and therefore, when we construct the external url, we should
ignore the ironic-proxy altogether.