Reduce container privilege
Many of the containers we run have to be run in privileged mode in order
to access a host volume - not only so they can write files as root, but
also to get an SELinux context that allows it.

However, some containers currently running as privileged do not require
this.