Remove metrics port from baremetal-operator
baremetal-operator isn't listening on port 60000, the metrics
listener is on port 8085 (the default), it dosn't need to be
defined in the pod as we arn't using it. Defining a HostPort
(even an unused one) will prevent the pod from being schedulable
if the port is already in use on the host.