OCPBUGS-23149, OCPBUGS-23131: fix IRONIC_EXTERNAL_URL_V6
IRONIC_EXTERNAL_URL_V6 is currently calculated incorrectly:

1) It uses the API VIP, but the Ironic's httpd is not proxied in the
   same way as Ironic itself and thus may not be available there.
2) It uses an external IP even when a provisioning network is enabled
   and VirtualMediaViaExternalNetwork is false.

This change fixes both issues, while also cleaning up the provisioning
utils to ensure all functions are dualstack-aware. The split between
the BMO and Metal3 pods is an implicit requirement of this change.

This change does NOT fix the fact that ICC is not configured correctly
for deploying IPv6-only nodes from a dualstack cluster.