Move parts of ICC config to a new secret
The assisted service needs to get the same configuration to embed IPA
into their images. Right now they embed CBO and ICC code to achieve
that. When we change anything on the CBO side (e.g. add ironic-proxy or
drop inspector), they need to update their code on all supported
releases.

This change adds a new secret with some of the IRONIC_* variables. It
will be used by the assisted service instead of CBO code directly.