Refactor Pull Secret application
There is no need for us to reproduce the logic of library-go's
ApplySecret(), which applies the changes iff the secret has changed. Our
local applySecret() with its custom callback is only needed when we want
to prevent updates in some circumstances where the secret data is
different. The error in the original code (prior to
80f5e6fe41f82ae06c607823f09d7d26cd447868) was that we passed
doNotUpdateData (which always returns false) as the condition.

This code was doing 3 Get()s where only 1 is required.