METAL-745: Make CBO aware of HostFirmwareComponents
This PR updates BMO to point to the latest version we have downstream,
the version includes changes that are necessary for
HostFirmwareComponents.

NOTE: with controller-runtime 0.16.3  a lot of things changed on how
we can specify NewCache, I'm attempting to simplify things by using
`Cache` instead of `NewCache`.

Also includes hostfirmwarecomponents in relatedObjects for the
clusteroperator controller.

Signed-off-by: Iury Gregory Melo Ferreira <imelofer@redhat.com>