Allow provisioningIP when network disabled
In 96b70d809bb608194407ebea334979a3e91a5a74, we decided to ignore the
ProvisioningIP setting when the provisioning network is disabled.  It
turns out that this has unintended consequences in specific scenarios.

Here, we are relaxing that requirement.  We pass the ProvisioningIP
value to the metal3 containers.  The original bug was related to DHCP
causing problems so we disable DHCP when the provisioning network is
disabled.