Allow provisioningIP when network disabled
In 96b70d8, we decided to ignore the ProvisioningIP setting when the
provisioning network is disabled. It turns out that this has unintended
consequences in specific scenarios.

Here, we are relaxing that requirement. We pass the ProvisioningIP value
to the metal3 containers. The original bug was related to DHCP causing
problems so we disable DHCP when the provisioning network is disabled.

We also take the opportunity to refactor a bit.  The provisioning
network CIDR only makes sense in the case of a Managed network (DHCP).
We separate those concerns while making sure to pass the CIDR
provisioning IP along when using DHCP.  The ironic-image strips it out
and the static-ip image depends on it.