Skip iso files in baremetal pod
Both the baremetal pod and the image customization pod run
machine-os-images as an init container to download some files.  This
patch changes the baremetal pod version to skip copying iso files.
These iso files are only needed in image customization and so we allow
image customization to acquire them when runs.